Chore: 예외처리 폴더 : Category => category 로 폴더명 변경 #27
