+++ plot-server/src/main/java/com/plot/plotserver/exception/category/CategoryAlreadyExistException.java
@@ -1,22 +1,20 @@
-package com.plot.plotserver.exception.Category;
+package com.plot.plotserver.exception.category;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
+
+
 @Getter
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class CategoryUpdateFailException extends RuntimeException {
+public class CategoryAlreadyExistException extends  RuntimeException{
 
     private String message;
 
-    public CategoryUpdateFailException(String message) {
+    public CategoryAlreadyExistException(String message) {
         super(message);
         this.message = message;
     }
 }
-
-
-
-
