+++ plot-server/src/main/java/com/plot/plotserver/exception/category/CategoryDeleteFailException.java
@@ -1,4 +1,4 @@
-package com.plot.plotserver.exception.Category;
+package com.plot.plotserver.exception.category;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
@@ -7,13 +7,18 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @Getter
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class CategorySavedFailException extends RuntimeException {
+public class CategoryDeleteFailException extends  RuntimeException{
 
     private String message;
 
-    public CategorySavedFailException(String message) {
+    public CategoryDeleteFailException(String message) {
         super(message);
         this.message = message;
     }
-
 }
+
+
+
+
+
+
