+++ plot-server/src/main/java/com/plot/plotserver/exception/category/CategorySavedFailException.java
@@ -1,4 +1,4 @@
-package com.plot.plotserver.exception.Category;
+package com.plot.plotserver.exception.category;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
@@ -7,18 +7,13 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @Getter
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class CategoryDeleteFailException extends  RuntimeException{
+public class CategorySavedFailException extends RuntimeException {
 
     private String message;
 
-    public CategoryDeleteFailException(String message) {
+    public CategorySavedFailException(String message) {
         super(message);
         this.message = message;
     }
+
 }
-
-
-
-
-
-
