+++ plot-server/src/main/java/com/plot/plotserver/exception/category/CategoryUpdateFailException.java
@@ -1,20 +1,22 @@
-package com.plot.plotserver.exception.Category;
+package com.plot.plotserver.exception.category;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
-
-
 @Getter
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class CategoryAlreadyExistException extends  RuntimeException{
+public class CategoryUpdateFailException extends RuntimeException {
 
     private String message;
 
-    public CategoryAlreadyExistException(String message) {
+    public CategoryUpdateFailException(String message) {
         super(message);
         this.message = message;
     }
 }
+
+
+
+
