+++ plot-server/src/main/java/com/plot/plotserver/service/CategoryService.java
@@ -5,10 +5,10 @@ import com.plot.plotserver.domain.Category;
 import com.plot.plotserver.domain.CategoryGroup;
 import com.plot.plotserver.dto.request.category.NewCategoryReqDto;
 import com.plot.plotserver.dto.request.category.UpdateCategoryReqDto;
-import com.plot.plotserver.exception.Category.CategoryAlreadyExistException;
-import com.plot.plotserver.exception.Category.CategoryDeleteFailException;
-import com.plot.plotserver.exception.Category.CategorySavedFailException;
-import com.plot.plotserver.exception.Category.CategoryUpdateFailException;
+import com.plot.plotserver.exception.category.CategoryAlreadyExistException;
+import com.plot.plotserver.exception.category.CategoryDeleteFailException;
+import com.plot.plotserver.exception.category.CategorySavedFailException;
+import com.plot.plotserver.exception.category.CategoryUpdateFailException;
 import com.plot.plotserver.repository.CategoryGroupRepository;
 import com.plot.plotserver.repository.CategoryRepository;
 import com.plot.plotserver.repository.TagCategoryRepository;
